moving pdstemplates.h contents into pdstemplates.c
